Activities and fragments moved to new UI package
