+++ app/src/main/AndroidManifest.xml
@@ -10,7 +10,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".SearchUsersActivity"
+            android:name=".UI.SearchUsersActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
@@ -20,13 +20,13 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".UserDetailActivity"
+            android:name=".UI.UserDetailActivity"
             android:label="@string/title_item_detail"
-            android:parentActivityName=".SearchUsersActivity"
+            android:parentActivityName=".UI.SearchUsersActivity"
             android:theme="@style/AppTheme.NoActionBar">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.mapelli.simone.githubclient.SearchUsersActivity" />
+                android:value="com.mapelli.simone.githubclient.UI.SearchUsersActivity" />
         </activity>
     </application>
