+++ app/src/main/res/layout/activity_user_detail.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".UserDetailActivity"
+    tools:context=".UI.UserDetailActivity"
     tools:ignore="MergeRootFrame">
