+++ app/src/main/res/layout/activity_user_list.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context=".SearchUsersActivity">
+    tools:context=".UI.SearchUsersActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
