+++ app/src/main/res/layout/fragment_repositories.xml
@@ -5,7 +5,7 @@
     android:id="@+id/constraintLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".RepositoriesFragment">
+    tools:context=".UI.RepositoriesFragment">
 
     <TextView
         android:id="@+id/repository_list_txt"
