+++ app/src/main/res/layout/user_list.xml
@@ -9,5 +9,5 @@
     android:layout_marginLeft="16dp"
     android:layout_marginRight="16dp"
     app:layoutManager="LinearLayoutManager"
-    tools:context=".SearchUsersActivity"
+    tools:context=".UI.SearchUsersActivity"
     tools:listitem="@layout/user_list_item" />
