+++ app/src/main/res/values/strings.xml
@@ -7,5 +7,5 @@
     <string name="loadMore_btn_title">Load More</string>
     <string name="profile_tab_title">Profile</string>
     <string name="repositories_tab_title">Repositories</string>
-    
+
 </resources>
